+++ hexbin/globals.h
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 #include <string.h>
 #ifdef BSD
 extern char *rindex();
@@ -18,10 +19,10 @@ typedef struct macheader {
 	char m_type[4];
 	char m_author[4];
 	short m_flags;
-	long m_datalen;
-	long m_rsrclen;
-	long m_createtime;
-	long m_modifytime;
+	int32_t m_datalen;
+	int32_t m_rsrclen;
+	int32_t m_createtime;
+	int32_t m_modifytime;
 } macheader;
 
 extern struct macheader mh;
